Dev (#4)
* add version, ready, live and request logging middlewares

* add Infrastructure project with all filters and middlewares

* add grpc proj and add startup filter for it

* add models, interceptors, http client and service and controllers

* update models and add restease project

* update resteasy project name

* remove httpclients proj

* remove directory

* Add domain adn test projects

* update domain models add UpdateStatus logic and test for it

* add repository interfaceses a and tests

* add mediatr, and use it in controllers

* update domain model and add some business logic, add command handlers and map it into controllers

* update controllers using block

* update tests for update order state logic

* add IAgregateRoot, change models, change updating order state logic, update tests

* remove merchrepository from mock

* remove merchrepository from extensions

* add changes into domain

* add application project and movecommands and queries there

* update docker-compose

* add migration base, add repository infrastructure, update docker and docker compose files

* Add repository instance, add unit of work and chacnge tracker, update handlers and host builder extension

* change domain model

* update tests and domain, add migrations and repositories

* change returning types, update queries params...

* add restease to client and ant and api projects

+++ OzonEdu.Merchandise.Domain/Models/Entity.cs
@@ -7,8 +7,8 @@ namespace OzonEdu.Merchandise.Domain.Models
     public abstract class Entity
     {
         int? _requestedHashCode;
-        int _Id;
-        public virtual int Id
+        long _Id;
+        public virtual long Id
         {
             get
             {
@@ -41,7 +41,7 @@ namespace OzonEdu.Merchandise.Domain.Models
 
         public bool IsTransient()
         {
-            return this.Id == default(Int32);
+            return this.Id == default(Int64);
         }
 
         public override bool Equals(object obj)
